requirements.txt
Jinja2>=3.1,<4
python-dotenv>=1.0,<2
websockets>=12,<16
cryptography>=42,<46
